Consolidate challenge API configuration module
Committed-By-Agent: codex
